+++ changelog.md
@@ -1,8 +1,19 @@
 # changelog
 
+## 1.6.0
+* `NEW` auto require local modules
+* `NEW` hover function by keyword `function`
+* `NEW` code action: swap params
+* `CHG` unbind the relative path between binaries and scripts
+
 ## 1.5.0
+`2020-12-5`
 * `NEW` setting `runtime.unicodeName`
+* `NEW` fully supports `---@generic T`
 * `FIX` [#274](https://github.com/sumneko/lua-language-server/issues/274)
+* `FIX` [#276](https://github.com/sumneko/lua-language-server/issues/276)
+* `FIX` [#279](https://github.com/sumneko/lua-language-server/issues/279)
+* `FIX` [#280](https://github.com/sumneko/lua-language-server/issues/280)
 
 ## 1.4.0
 `2020-12-3`